comparison of integer expressions of different signedness fix

+++ 16-binary_tree_is_perfect.c
@@ -47,6 +47,5 @@ int binary_tree_is_perfect(const binary_tree_t *tree)
 	size_t height = binary_tree_height(tree);
 	size_t size = binary_tree_size(tree);
 
-	return (size == (1 << height) - 1);
+	return (size == ((size_t)1 << height) - 1);
 }
-
